+++ resources/views/components/action-button.blade.php
@@ -1,1 +1,1 @@
-<a href="{{ $route }}" class="btn btn-primary btn-lg d-md-none floating-action-button">@icon({{ $icon }})</a>
+<a href="{{ $route }}" class="btn btn-primary btn-lg d-md-none floating-action-button {{ $fab_class ?? '' }}">@icon({{ $icon }})</a>